services/vector-search/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.3
httpx==0.26.0
redis==5.0.1
